* clean code : removed useless "using", set all grid unit from float* to int*

+++ Assets/Scripts/GameElements/InputPlaceholder.cs
@@ -8,11 +8,11 @@ namespace Game.GameElements
 
 public class InputPlaceholder : MonoBehaviour
 {
-	
+
 	// ACCESSORS
 
-	public float2 GridPosition { get; set; }
-	
+	public int2 GridPosition { get; set; }
+
 }
 
 }
